+++ examples/HttpStatusCode/App.tsx
@@ -39,7 +39,7 @@ class Form extends React.Component<Props, State> {
   async handleSubmit(e: React.FormEvent<HTMLFormElement>) {
     e.preventDefault();
 
-    await this.form.validateFields();
+    await this.form.validateForm();
     if (this.form.isValid()) {
       alert(`Valid form\n\nthis.state =\n${JSON.stringify(this.state, null, 2)}`);
     } else {