+++ examples/MDN_Form_validation/App.tsx
@@ -24,7 +24,7 @@ class Form extends React.Component {
   async handleSubmit(e: React.FormEvent<HTMLFormElement>) {
     e.preventDefault();
 
-    await this.form.validateFields();
+    await this.form.validateForm();
     if (this.form.isValid()) {
       alert('Valid form');
     } else {