+++ examples/WizardForm/WizardFormPage1.tsx
@@ -30,7 +30,7 @@ class WizardFormPage1 extends React.Component<Props> {
   async handleSubmit(e: React.FormEvent<HTMLFormElement>) {
     e.preventDefault();
 
-    await this.form.validateFields();
+    await this.form.validateForm();
     if (this.form.isValid()) {
       this.props.nextPage();
     }