+++ examples/WizardForm/WizardFormPage2.tsx
@@ -33,7 +33,7 @@ class WizardFormPage2 extends React.Component<Props> {
   async handleSubmit(e: React.FormEvent<HTMLFormElement>) {
     e.preventDefault();
 
-    await this.form.validateFields();
+    await this.form.validateForm();
     if (this.form.isValid()) {
       this.props.nextPage();
     }